operations: List...
Adjusted the Reduce class to handle a list of operations rather than a dict of path and methods.

A string or pattern can be used in the list to match against operationIds, or a tuple with a string or pattern as the first element representing the path and a list of strings representing HTTP verbs as the second element will match against the path/methods.

Also, in any case that a path is not culled, the non-operation keys are kept intact to maintain defaults and overrides.

The MSGraph test was updated to accommodate the annotation change as well as to handle a few validation fixes.

(cherry picked from commit 21b8dfb663451ebf71e1b1acb8333fc58e3b949e)